JavaFX textarea, textfield and checkbox elements from tabs have been fixed

+++ src/main/java/com/henrikbeck95/swordle/dictionary/DictionaryFile.java
@@ -10,9 +10,9 @@ import java.nio.file.Paths;
  */
 
 public class DictionaryFile {
-	public static int getAmountOfLines(String filePath) {
+	public static long getAmountOfLines(String filePath) {
 		Path path = Paths.get(filePath);
-		int lines = 0;
+		long lines = 0;
 
 		try {
 			lines = (int) Files.lines(path).count();
